feat: attribute outbound Databricks requests to AppKit consistently
Only the service and user WorkspaceClients carried the AppKit User-Agent
(@databricks/appkit/<version>). The cache client sent unknown/0.0.0, and
the two raw-fetch sites (files upload, MCP connector) sent no AppKit
User-Agent at all, since they bypass the SDK apiClient.

Extract getClientOptions() into a shared context/client-options module
(adding APPKIT_USER_AGENT for fetch sites with no apiClient), pass it to
the cache client, and stamp the User-Agent on the files upload fetch (via
apiClient.userAgent()) and both MCP fetches.
