+++ packages/appkit/src/connectors/mcp/client.ts
@@ -23,6 +23,7 @@
  * transport.
  */
 import type { AgentToolDefinition } from "shared";
+import { APPKIT_USER_AGENT } from "../../context/client-options";
 import { createLogger } from "../../logging/logger";
 import {
   assertResolvedHostSafe,
@@ -423,6 +424,9 @@ export class AppKitMcpClient {
 
     const authHeaders = await this.resolveAuthHeaders(options);
     const headers: Record<string, string> = {
+      // Raw fetch bypasses the SDK's apiClient; stamp the AppKit User-Agent so
+      // MCP traffic is attributed to AppKit.
+      "User-Agent": APPKIT_USER_AGENT,
       "Content-Type": "application/json",
       Accept: "application/json, text/event-stream",
       ...authHeaders,
@@ -497,6 +501,9 @@ export class AppKitMcpClient {
 
     const authHeaders = await this.resolveAuthHeaders(options);
     const headers: Record<string, string> = {
+      // Raw fetch bypasses the SDK's apiClient; stamp the AppKit User-Agent so
+      // MCP traffic is attributed to AppKit.
+      "User-Agent": APPKIT_USER_AGENT,
       "Content-Type": "application/json",
       Accept: "application/json, text/event-stream",
       ...authHeaders,
